--- Assets/#0 Scenes/main.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ee7843a3c598a40e290bdc59abfd5d7c
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
