+++ Assets/#2 Prefabs/Player/player1.prefab
@@ -374,7 +374,7 @@ GameObject:
   - component: {fileID: 8523253409257125381}
   - component: {fileID: 826423447707642263}
   - component: {fileID: 8196518570531318362}
-  - component: {fileID: 1052625459820057633}
+  - component: {fileID: 8963870929831639437}
   m_Layer: 0
   m_Name: CheckEnemy
   m_TagString: Untagged
@@ -477,7 +477,7 @@ PolygonCollider2D:
     - - {x: 0, y: 0.5773587}
       - {x: -0.5, y: -0.28866667}
       - {x: 0.5, y: -0.28866667}
---- !u!114 &1052625459820057633
+--- !u!114 &8963870929831639437
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -486,10 +486,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 6810298835053482159}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fe8790bd7046fa742886130f8d4a617d, type: 3}
+  m_Script: {fileID: 11500000, guid: 137399c0d6413d74bb5f2cb890b77112, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   Dash_Enemy: []
+  IsChecked: 0
 --- !u!1 &7341301570535230410
 GameObject:
   m_ObjectHideFlags: 0
